Add minimal web app
Use flask to render web app containing a single button sending a toggle
command via MQTT. The actual MQTT request is send from a server, since
the long term idea is not to have a manual control, but to automate the
control from the server. The web interface is just a dummy setup to
demonstrate the information flow.